requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.4.0
django-cors-headers>=4.6.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
celery>=5.4.0
redis>=5.2.1
spacy>=3.8.2
nltk>=3.9.1
scikit-learn>=1.5.2
pandas>=2.2.3
numpy>=2.1.3
joblib>=1.4.2
pdfplumber>=0.11.5
python-docx>=1.1.2
drf-spectacular>=0.28.0
